requirements.txt
cairocffi==1.5.0
CairoSVG==2.7.0
cffi==1.15.1
contourpy==1.0.7
csdt-stl-converter==1.2.1
csdt-stl-tools==0.5.4
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
fonttools==4.39.3
importlib-resources==5.12.0
imutils==0.5.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.7.1
numpy==1.23.5
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.5.0
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
reportlab==3.6.12
scipy==1.10.1
six==1.16.0
svglib==1.5.1
tinycss2==1.2.1
trimesh==3.21.3
webencodings==0.5.1
zipp==3.15.0
